refactor filemanager.py for code climate
